pull out the hide modal logic, use it in two places
