+++ packages/compass-e2e-tests/helpers/commands/index.ts
@@ -59,5 +59,6 @@ export * from './create-index';
 export * from './drop-index';
 export * from './hide-index';
 export * from './unhide-index';
+export * from './hide-visible-modal';
 export * from './hide-visible-toasts';
 export * from './sidebar-collection';
